% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basin_hillslope_stats.R
\name{basin_hillslope_stats}
\alias{basin_hillslope_stats}
\title{basin_hillslope_stats}
\usage{
basin_hillslope_stats(worldfile, outfile = NULL, LPC_vegnames = T)
}
\arguments{
\item{worldfile}{Source worldfile}

\item{outfile}{Output text file containing formatted tables.}

\item{LPC_vegnames}{TRUE/FALSE Should the veg_parm_IDs of 1,5,3,4 be automatically renamed to tree, shrub, herb, and bare (other).}
}
\description{
Using a worldfile, generates stats at the basin and hillslope level, including slope, elevation, and distribution of veg cover types. Returns a list of tables in R, and optionally outputs a text file of tables.
}
