% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_meta.R
\name{build_meta}
\alias{build_meta}
\title{build_meta}
\usage{
build_meta(
  name,
  world = NULL,
  flow = NULL,
  template,
  map_dir,
  cf_maps,
  streams = NULL,
  roads = NULL,
  impervious = NULL,
  roofs = NULL,
  asp_rule = NULL
)
}
\arguments{
\item{name}{Name and path for metadata. ".meta" will be appended automatically.}

\item{world}{worldfile}

\item{flow}{flowtable}

\item{template}{template}

\item{map_dir}{dir for spatial input}

\item{cf_maps}{additional maps input for flownet creation}

\item{streams}{stream map}

\item{roads}{roads map}

\item{impervious}{impervious map}

\item{roofs}{roofs map}

\item{asp_rule}{aspatial rule file}
}
\value{
Writes metadata with name and path indicated in name
}
\description{
Create documentation/metadata for RHESSys Preprocessing. 2/27/18
}
\author{
Will Burke
}
