% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_redefine_template.R
\name{build_redefine_template}
\alias{build_redefine_template}
\title{Build Redefine Template}
\usage{
build_redefine_template(template, name, overwrite)
}
\arguments{
\item{template}{Original template file used to generate your worldfile for RHESSys.}

\item{name}{Name of output redefine template to be created}

\item{overwrite}{Overwrite existing redefine template? FALSE is default and prompts a menu if worldfile already exists.}
}
\description{
Builds a template file for use in a redefine world. Pair with world_redefine.
}
\author{
Will Burke
}
