% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_landscape_rasters_by_basin.R
\name{clip_landscape_rasters_by_basin}
\alias{clip_landscape_rasters_by_basin}
\title{Clip landscape rasters by basin}
\usage{
clip_landscape_rasters_by_basin(
  work_folder,
  file_identifier = "_landscape",
  return_rasters = FALSE
)
}
\arguments{
\item{work_folder}{Path to folder containing inputs/outputs. Inputs should
include rasters to be clipped.}

\item{file_identifier}{Syntax in file name that identifies a raster for
clipping. The file_identifier is removed from the outout file name.}

\item{Option}{to return clipped rasters as list of R objects in addition to
exporting to file.}
}
\description{
This function clips landscape rasters by the basin extent for use in
`RHESSysPreprocess.
}
\author{
Ryan Bart
}
