% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clump_bycategory.R
\name{clump_bycategory}
\alias{clump_bycategory}
\title{clump_bycategory}
\usage{
clump_bycategory(basemap, directions = 8)
}
\arguments{
\item{basemap}{a raster map values should be integer (e.g use round) for clumping}

\item{direction}{(default 8) direction for clumping}
}
\value{
raster with same coordinates/dimensions as basemap with clump ids
}
\description{
create patches with unique ids for spatially contigous areas with the same value
}
