% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_flowtable.R
\name{convert_flowtable}
\alias{convert_flowtable}
\title{convert_flowtable}
\usage{
convert_flowtable(input_file, output_file, make_stream = 4)
}
\arguments{
\item{input_file}{Flow table to be read in.}

\item{output_file}{Name for output file to be created.}

\item{make_stream}{The maximum distance (cell lengths) away from an existing stream that a patch can be automatically coerced to be a stream.
Setting to TRUE will include patches at any distance. This is needed for hillslope parallelization, as all hillslopes must have an outlet stream patch.
 Default is 4.}
}
\description{
Convert a flowtable to hillslope parallelized format
}
\author{
William Burke
}
