% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateFlownet.R
\name{CreateFlownet}
\alias{CreateFlownet}
\title{CreateFlownet}
\usage{
CreateFlownet(flownet_name, readin = NULL, type = "raster",
  typepars = NULL, asp_list = NULL, streams = NULL,
  overwrite = FALSE, roads = NULL, road_width = NULL,
  impervious = NULL, roofs = NULL, wrapper = FALSE,
  parallel = TRUE, make_stream = 4, d4 = FALSE)
}
\arguments{
\item{flownet_name}{The name of the flow network file to be created.  Will be coerced to have ".flow" extension if not already present.}

\item{readin}{readin indicates which maps to be used. If CreateFlowmet.R is run it's own, this should point to the template. Otherwise,
if run inside of RHESSysPreprocess, readin will use the map data from world_gen.R, Streams map, and other optional maps, still need to
be specified.}

\item{type}{Input file type to be used. Default is raster. "Raster" type will use rasters
in GeoTiff or equivalent format (see Raster package), with file names  matching those indicated in the template.
ASCII is supported, but 0's cannot be used as values for data. "GRASS" will attempt to autodetect the version of
GRASS GIS being used (6.x or 7.x).  GRASS GIS type can also be set explicitly to "GRASS6" or "GRASS7".}

\item{typepars}{Parameters needed based on input data type used. If using raster type, typepars should be a string
indicating the path to a folder containing the raster files that are referenced by the template.
For GRASS GIS type, typepars is a vector of 5 character strings. GRASS GIS parameters: gisBase, home, gisDbase, location, mapset.
Example parameters are included in an example script included in this package. See initGRASS help
for more info on parameters.}

\item{streams}{Streams map to be used in building the flowtable.}

\item{overwrite}{Overwrite existing worldfile. FALSE is default and prompts a menu if worldfile already exists.}

\item{roads}{Roads map, an optional input for flowtable creation.}

\item{impervious}{Impervious map, an optional input for flowtable creation.}

\item{roofs}{Roofs map, an optional input for flowtable creation.}

\item{parallel}{TRUE/FALSE flag to build a flowtable for use in the hilllslope parallelized version of RHESSys. Console may output warnings of
automated actions taken to make hillslope parallelization possible, or errors indicating fatal problems in hillslope parallelization.}

\item{make_stream}{The maximum distance (cell lengths) away from an existing stream that a patch can be automatically coerced to be a stream.
Setting to TRUE will include patches at any distance. This is needed for hillslope parallelization, as all hillslopes must have an outlet stream patch.
 Default is 4.}

\item{d4}{TRUE/FALSE flag to determine the logic used when finding neighbors in flow table creation. FALSE uses d8 routing, looking at all eight
neighboring cells. TRUE uses d4 routing, looking at only cardinal directions, not diagonals.}
}
\description{
Creates the flow networkd table used by RHESSys
}
\author{
Will Burke
}
