% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_terrain_variables.R
\name{download_terrain_variables}
\alias{download_terrain_variables}
\title{Download terrain (DEM and NLCD) variables for U.S.}
\usage{
download_terrain_variables(
  bbox,
  proj_epsg,
  data_source,
  variable,
  label,
  nlcd_year = NULL,
  res = NULL,
  res_example = NULL,
  output_file
)
}
\arguments{
\item{bbox}{A spatraster that defines the Area of Interest. Coordinates
should be in long/lat. Needed functionality: a bbox vector in the form of
c(........).}

\item{proj_epsg}{The UTM EPSG for reprojecting the terrain raster.}

\item{data_source}{Source of data. Current option is "FedData".}

\item{variable}{Name of the desired variable. Current options include "dem",
and "nlcd".}

\item{label}{The name of the study area.}

\item{nlcd_year}{Year of nlcd data. Acceptable values are 2019 (default),
2016, 2011, 2008, 2006, 2004, and 2001.}

\item{res}{Desired resolution in meters.}

\item{res_example}{An spatraster with desired resolution for the downloaded
data to be resampled to. In many cases, this will be same raster as bbox.
If both `res` and `res_example` are not null, `res_example` takes
precedent.}

\item{output_file}{Path and file name of raster to be saved.}
}
\description{
This function will download a National Elevation Dataset (NED) DEM or
National Land Cover Database (NLCD) raster for an area of interest, reproject
the raster, and adjust spatial resolution.
}
\author{
Ryan Bart
}
