% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_make_net_2d.R
\name{file_make_net_2d}
\alias{file_make_net_2d}
\title{file_make_net_2d}
\usage{
file_make_net_2d(flw_file)
}
\arguments{
\item{flw_file}{Flow table to generate diagram}
}
\description{
Reads in flow table and makes 2d diagram of flow routing
}
\author{
Daniel Nash
}
