% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_make_net_3d.R
\name{file_make_net_3d}
\alias{file_make_net_3d}
\title{file_make_net_3d}
\usage{
file_make_net_3d(flw_file)
}
\arguments{
\item{flw_file}{Flow table to generate diagram}
}
\description{
Reads in flow table and makes 3d diagram of flow routing
}
\author{
Daniel Nash
}
