% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_basin_raster.R
\name{make_basin_raster}
\alias{make_basin_raster}
\title{Make a basin raster}
\usage{
make_basin_raster(
  work_folder,
  pour_point_file = NULL,
  pour_point_longlat = NULL,
  proj_epsg = NULL,
  streams = "streams_landscape.tif",
  d8_pointer = "d8_pointer_landscape.tif",
  snap_dist,
  save_map = TRUE
)
}
\arguments{
\item{work_folder}{Path to folder containing inputs/outputs}

\item{pour_point_file}{Filename&extension (if located in work_folder), or
path&filename&extension, of a shapefile containing a point designating the
basin outlet. Projection of shapefile should be same as `streams` and
`d8_pointer` This argument overrides `pour_point_longlat`.}

\item{pour_point_longlat}{Vector containing the longitude and latitude for a
point designating the basin outlet. The coordinate will be reprojected to
the `proj_epsg` projection and then a shapefile will be generated, which
will subsequently be used as a direct input into Whitebox function. The new
shapefile will be written to `work_folder`. This argument will be ignored
if `pour_point_file` is not NULL.}

\item{proj_epsg}{EPSG of project that `pour_point_longlat` will be converted
to. Should be a UTM zone.}

\item{streams}{Filename&extension (if located in work_folder), or
path&filename&extension, of the streams raster to be inputted.}

\item{d8_pointer}{Filename&extension (if located in work_folder), or
path&filename&extension, of the D8 pointer raster to be inputted.}

\item{snap_dist}{Maximum snap distance in map units (Meters for UTM).}

\item{save_map}{Option to save a map showing pour point adjustment on stream
network.}
}
\description{
This function generates a basin raster by 1) adjusting a pour point (aka
outlet) to the stream network, and 2) cutting out the corresponding basin.
}
\section{Note}{
 WhiteboxTools Open Core (https://www.whiteboxgeo.com/) must be
  installed on your computer to run this function. The R package whitebox is
  an R wrapper that interfaces with Whitebox.

  Several improvements can be made to this function. 1) May be
  able to get rid of `proj_epsg` and take desired projection directly from
  `streams` and `d8_pointer`. 2) It would probably be better to generate
  shiny app or leaflet map to identify outlet.

  Links to embedded functions
  https://www.whiteboxgeo.com/manual/wbt_book/available_tools/hydrological_analysis.html#JensonSnapPourPoints
  https://www.whiteboxgeo.com/manual/wbt_book/available_tools/geomorphometric_analysis.html#Watershed
}

\author{
Ryan Bart
}
