% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_flow_list.R
\name{make_flow_list}
\alias{make_flow_list}
\title{patch_data_analysis}
\usage{
make_flow_list(
  raw_patch_data,
  raw_patch_elevation_data,
  raw_hill_data,
  raw_basin_data,
  raw_zone_data,
  raw_slope_data,
  raw_stream_data,
  raw_road_data,
  cell_length,
  road_width = NULL,
  parallel,
  make_stream,
  skip_hillslope_check
)
}
\arguments{
\item{raw_patch_data}{patch matrix}

\item{raw_patch_elevation_data}{DEM matrix}

\item{raw_hill_data}{hillslope matrix}

\item{raw_basin_data}{basin matrix}

\item{raw_zone_data}{zone matrix}

\item{raw_slope_data}{slope matrix}

\item{raw_stream_data}{stream map matrix}

\item{raw_road_data}{road map matrix}

\item{cell_length}{cell length in meters}

\item{road_width}{>0, defaults to 1.}

\item{parallel}{TRUE/FALSE flag to build a flowtable for use in the hilllslope parallelized version of RHESSys. Console may output warnings of
automated actions taken to make hillslope parallelization possible, or errors indicating fatal problems in hillslope parallelization.}

\item{make_stream}{The maximum distance (cell lengths) away from an existing stream that a patch can be automatically coerced to be a stream.
Setting to TRUE will include patches at any distance. This is needed for hillslope parallelization, as all hillslopes must have an outlet stream patch.
Default is 4.}

\item{skip_hillslope_check}{TRUE/FALSE to skip the recursive check for segmented hillslopes. Segmented hillslopes will break the routing, but the
recursive check can trigger various recursion protections when hillslopes are large.}
}
\description{
This function is the main component of the larger create_flownet function.
Function finds neighbors, calculates flow, accounts for pits, roads, streams etc.
cell_length is size of cell in meters.
}
\author{
Daniel Nash

Will Burke
}
