% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_landscape_rasters.R
\name{make_landscape_rasters}
\alias{make_landscape_rasters}
\title{Make landscape rasters for watershed processing}
\usage{
make_landscape_rasters(
  dem,
  work_folder,
  stream_threshold,
  patch_method = "simple"
)
}
\arguments{
\item{dem}{Filename&extension (if located in work_folder), or
path&filename&extension, of the DEM to be inputted. DEM must be a file on
hard drive, not an object within R. The function assumes that the DEM is
already at the desired resolution, extent and projection. Projection should
be in UTM, as some of the whitebox computations assume that pixel
dimensions are in meters.}

\item{work_folder}{Path to folder containing inputs/outputs}

\item{stream_threshold}{Threshold of flow accumulation for designating a
patch as stream.}

\item{patch_method}{Options for generating patch raster. Options include
'simple' (default).}
}
\description{
Using a DEM input, this function generates aspect, slope, horizon (east and
west), d8 pointer, d8 flow accumulation, stream network, subbasin, and patch
rasters.
}
\section{Note}{
 WhiteboxTools Open Core (https://www.whiteboxgeo.com/) must be
  installed on your computer to run this function. The R package whitebox is
  an R wrapper that interfaces with Whitebox.

  The raster outputs from this function include filenames indicating the type
  of data plus an appended "_landscape". The "_landscape" signifies that the
  raster values encompasses the entire bounding box, as opposed to simply the
  basin, which will be generated in subsequent functions.

  Links to embedded whitebox functions
  https://www.whiteboxgeo.com/manual/wbt_book/available_tools/geomorphometric_analysis.html#Aspect
  https://www.whiteboxgeo.com/manual/wbt_book/available_tools/geomorphometric_analysis.html#Slope
  https://www.whiteboxgeo.com/manual/wbt_book/available_tools/geomorphometric_analysis.html#HorizonAngle
  https://www.whiteboxgeo.com/manual/wbt_book/available_tools/hydrological_analysis.html#BreachDepressions
  https://www.whiteboxgeo.com/manual/wbt_book/available_tools/hydrological_analysis.html#D8Pointer
  https://www.whiteboxgeo.com/manual/wbt_book/available_tools/hydrological_analysis.html#D8FlowAccumulation
  https://www.whiteboxgeo.com/manual/wbt_book/available_tools/stream_network_analysis.html#ExtractStreams
  https://www.whiteboxgeo.com/manual/wbt_book/available_tools/hydrological_analysis.html#Subbasins
}

\author{
Ryan Bart
}
