% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_net_3d_pic.R
\name{make_net_3d_pic}
\alias{make_net_3d_pic}
\title{make_net_3d_pic}
\usage{
make_net_3d_pic(flw)
}
\arguments{
\item{flw}{Flow table list}
}
\description{
Input flow table to create a 3d diagram of flow routing
}
\author{
Daniel Nash
}
