% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscale_flow.R
\name{multiscale_flow}
\alias{multiscale_flow}
\title{multiscale_flow}
\usage{
multiscale_flow(CF1, asp_map, patch_map, asp_list)
}
\arguments{
\item{CF1}{Flowtable list object created from make_flow_list}

\item{asp_map}{Map of the aspatial rules}

\item{patch_map}{Map of the patches}

\item{asp_list}{List of aspatial rules}
}
\description{
Modifies an existing flowtable (R list) to work with multisacle routing.
}
