<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_flow.R
\name{read_in_flow}
\alias{read_in_flow}
\title{read_in_flow2}
\usage{
read_in_flow(input_file, df = FALSE)
}
\arguments{
\item{input_file}{Flow table to be read in.}

\item{df}{should a data frame also be output?}
}
\description{
Read in flow table file and return a R list. Should be faster than the other one.
}
\author{
William Burke
}
=======
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_flow.R
\name{read_in_flow}
\alias{read_in_flow}
\title{read_in_flow2}
\usage{
read_in_flow(input_file, df = FALSE)
}
\arguments{
\item{input_file}{Flow table to be read in.}

\item{df}{should a data frame also be output?}
}
\description{
Read in flow table file and return a R list. Should be faster than the other one.
}
\author{
William Burke
}
>>>>>>> 2755b10729a3f330336dcb9ff5558312c4f4ba5f
