% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_flow2.R
\name{read_in_flow2}
\alias{read_in_flow2}
\title{read_in_flow2}
\usage{
read_in_flow2(input_file)
}
\arguments{
\item{input_file}{Flow table to be read in.}
}
\description{
Read in flow table file and return a R list. Should be faster than the other one.
}
\author{
William Burke
}
