% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHESSysPreprocess.R
\name{RHESSysPreprocess}
\alias{RHESSysPreprocess}
\title{RHESSysPreprocess}
\usage{
RHESSysPreprocess(
  template,
  name,
  map_dir = NULL,
  typepars = NULL,
  streams = NULL,
  overwrite = FALSE,
  roads = NULL,
  impervious = NULL,
  roofs = NULL,
  header = FALSE,
  meta = FALSE,
  asprules = NULL,
  unique_strata_ID = TRUE,
  seq_patch_IDs = FALSE,
  output_patch_map = FALSE,
  fire_grid_out = FALSE,
  parallel = TRUE,
  make_stream = 4,
  skip_hillslope_check = FALSE,
  convert_aspect = TRUE,
  wrapper = TRUE
)
}
\arguments{
\item{template}{Template file used to generate worldfile for RHESSys. Generic strucutre is:
<state variable> <operator> <value/map>. Levels are difined by lines led by "_", structured
<levelname> <map> <count>. Whitespace and tabs are ignored.  Maps referred to must be supplied
by your chosen method of data input(raster).}

\item{name}{The base name (and potentially, path as well) to be used for your ouput files.
This will create a world file called "<name>.world", and a flow table called "<name>.flow".}

\item{map_dir}{Character string indicating the path to a folder containing the raster files that are referenced by the template.}

\item{typepars}{DEPRECATED arg, now use 'map_dir' to indicate the directory where your raster maps are located.}

\item{streams}{Streams map to be used in building the flowtable.}

\item{overwrite}{Overwrite existing worldfile. FALSE is default and prompts a menu if worldfile already exists.}

\item{roads}{Roads map, an optional input for flowtable creation.}

\item{impervious}{Impervious map, an optional input for flowtable creation.}

\item{roofs}{Roofs map, an optional input for flowtable creation.}

\item{header}{TRUE/FALSE flag for the creation of a header file. Will have same name (and location) as "name" argument, but with ".hdr" suffix.}

\item{meta}{TRUE/FALSE flag for the creation of a metadata file. Still in dev.}

\item{asprules}{The path and filename to the rules file.  Using this argument enables aspatial patches.}

\item{unique_strata_ID}{Takes input map or value for canopy strata ID and appends either a 1 or 2 dpending on which canopy it is. Defaults to TRUE.}

\item{seq_patch_IDs}{TRUE/FALSE should patch map IDs be overwritten with sequential integers.}

\item{output_patch_map}{TRUE/FALSE should the new patch map with sequential IDs be output to file.}

\item{fire_grid_out}{TRUE/FALSE should the writefire_grids headerless ascii grid outputs of DEM, patch, zone, and hillslope be output. An additional fire grid info
file will be created specifying the ncol/nrow and other header info. Will use the input name as the base prefix for the grid files.}

\item{parallel}{TRUE/FALSE flag to build a flowtable for use in the hilllslope parallelized version of RHESSys. Console may output warnings of
automated actions taken to make hillslope parallelization possible, or errors indicating fatal problems in hillslope parallelization.}

\item{make_stream}{The maximum distance (cell lengths) away from an existing stream that a patch can be automatically coerced to be a stream.
Setting to TRUE will include patches at any distance. This is needed for hillslope parallelization, as all hillslopes must have an outlet stream patch.
Default is 4.}

\item{skip_hillslope_check}{TRUE/FALSE to skip the recursive check for segmented hillslopes. Segmented hillslopes will break the routing, but the
recursive check can trigger various recursion protections when hillslopes are large.}

\item{convert_aspect}{TRUE/FALSE if the input aspect map data should be converted fom GRASS GIS format (CCW EAST==0) to the 'normal'
format needed for RHESSys (CW Noth==0). Defaults to TRUE.}

\item{wrapper}{internal argument to track if being run as all-in-one}
}
\description{
Preforms RHESSys Preprocessing, comprised of two main steps: generating a world file, and generating a flow table.
World file is generated via world_gen.R, and flow table is generated via create_flownet.R.
}
\seealso{
\code{\link[RHESSysIOinR]{run_rhessys}}
}
\author{
Will Burke
}
