% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_input_gen.R
\name{spatial_input_gen}
\alias{spatial_input_gen}
\title{spatial_input_gen}
\usage{
spatial_input_gen(
  name,
  DEM,
  GRASS_path,
  threshold,
  basin = NULL,
  patch = NULL,
  define_watershed = NULL,
  easting = NULL,
  northing = NULL,
  gisDbase = NULL,
  location = NULL,
  mapset = NULL
)
}
\arguments{
\item{basin}{The area defining your basin/watershed of interest. Assumes it's a raster}

\item{Name}{Name of your project}
}
\description{
Generate the spatial inputs for RHESSys and RHESSysPreprocessing
}
\author{
Will Burke
}
