% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_read.R
\name{template_read}
\alias{template_read}
\title{template_read}
\usage{
template_read(template)
}
\arguments{
\item{template}{template name and path to be read in}

\item{asp_check}{flag for if aspatial patches are being used/if the asp_rule map needs to be identified.}
}
\value{
Returns a list containing (in order): template in list form, variable name vector, index of levels,
index of variables, full table of maps and associated variables
}
\description{
Reads in a RHESSys template file, produces a clean list, variable name vector, indices of levels and variables
}
\author{
Will Burke
}
