% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_parallel.R
\name{test_parallel}
\alias{test_parallel}
\title{test_parallel}
\usage{
test_parallel(flw)
}
\arguments{
\item{flw}{List version of flow table, either produced via patch_data_analysis.R or read_in_flow2.R}
}
\description{
Checks for flow across hillslopes, and returns a table indicating patches that flow across hillslopes, 
how many patches they flow to, and the combined gamma of that patch going across hillslopes
}
\author{
Will Burke
}
