% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_template.R
\name{update_template}
\alias{update_template}
\title{update_template}
\usage{
update_template(
  template,
  out_file,
  vars,
  values,
  operator = NULL,
  level_names = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{template}{Name and path of templatefile to be created.}

\item{out_file}{Destination file to write new template}

\item{vars}{State variable(s) to be modified}

\item{values}{Values to replace for `vars`. N args must equal that of vars or be 1 (will be used for all vars).}

\item{operator}{Operator to be (optionally) replaced. If  not included the existing operator (e.g. 'value') will be left.}

\item{level_names}{Not implamented yet - for rare cases where you only want to replace 1 occurance of a state var (e.g. 'z' at zone but not patch ele)}

\item{overwrite}{TRUE/FALSE if input file should be overwriten}
}
\description{
Read and modify an existing template
}
\author{
Will Burke
}
