% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fire_grids.R
\name{write_fire_grids}
\alias{write_fire_grids}
\title{Write fire grids}
\usage{
write_fire_grids(name, template, map_dir, seq_patch_IDs = FALSE)
}
\arguments{
\item{name}{Basename for the output fire grids. This will be the fire grid prefix given to rhessys to specify the fire grid files.}

\item{template}{Input template file referencing the DEM, patch, zone, and hillslope spatial inputs.}

\item{map_dir}{Directory containing the input maps (referenced as 'typepars' elsewhere)}

\item{seq_patch_IDs}{TRUE/FALSE should sequential patch IDs be created (only needed if they were when worldfile was created)}
}
\description{
Writes ascii formatted, headerless, fire grid files for use with the RHESSys integrated WMFire model. Also outputs a info file
containing the grid dimensions and cell size.
}
\author{
Will Burke
}
