% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspatial_patches.R
\name{aspatial_patches}
\alias{aspatial_patches}
\title{Aspatial Patch Generation}
\usage{
aspatial_patches(asprules, asp_mapdata)
}
\description{
Aspatial Patch Generation
}
\author{
Will Burke
}
