% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_meta.R
\name{build_meta}
\alias{build_meta}
\title{build_meta}
\usage{
build_meta(name, template, worldfile, flownet, type, typepars, cf_maps, streams,
  roads = NULL, impervious = NULL, roofs = NULL)
}
\arguments{
\item{name}{Name and path for metadata. ".meta" will be appended automatically.}
}
\value{
Writes metadata with name and path indicated in name
}
\description{
Create documentation/metadata for RHESSys Preprocessing. 2/27/18
}
\author{
Will Burke
}
