% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_redefine_template.R
\name{build_redefine_template}
\alias{build_redefine_template}
\title{Build Redefine Template}
\usage{
build_redefine_template(template, name, replace, overwrite)
}
\arguments{
\item{template}{Original template file used to generate your worldfile for RHESSys.}

\item{name}{Name of output redefine template to be created}

\item{replace}{DOESNT WORK YET -Indicates the variable-method-values to replace. All others will be given values of "-9999"}

\item{overwrite}{Overwrite existing redefine template? FALSE is default and prompts a menu if worldfile already exists.}
}
\description{
Builds a template file for use in a redefine world.  Allows for specification of which variables to override and replace
with an absolute or multiplier value. Should be paired with world_redefine function (which is essentially world_gen), to
create the actual redefine worldfile.
}
\author{
Will Burke
}
