% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateFlownet.R
\name{CreateFlownet}
\alias{CreateFlownet}
\title{CreateFlownet}
\usage{
CreateFlownet(cfname, readin = NULL, type = "raster", typepars = NULL,
  asp_list = NULL, streams = NULL, overwrite = FALSE, roads = NULL,
  road_width = NULL, impervious = NULL, roofs = NULL, wrapper = FALSE,
  parallel = FALSE, d4 = FALSE)
}
\arguments{
\item{cfname}{The name of the flow network file to be created.  Will be coerced to have ".flow" extension if not already present.}

\item{readin}{readin indicates which maps to be used. If CreateFlowmet.R is run it's own, this should point to the template. Otherwise,
if run inside of RHESSysPreprocess, readin will use the map data from world_gen.R, Streams map, and other optional maps, still need to
be specified.}

\item{type}{Input file type to be used. Default is raster. "Raster" type will use rasters
in GeoTiff or equivalent format (see Raster package), with file names  matching those indicated in the template.
ASCII is supported, but 0's cannot be used as values for data. "GRASS" will attempt to autodetect the version of
GRASS GIS being used (6.x or 7.x).  GRASS GIS type can also be set explicitly to "GRASS6" or "GRASS7".}

\item{typepars}{Parameters needed based on input data type used. If using raster type, typepars should be a string
indicating the path to a folder containing the raster files that are referenced by the template.
For GRASS GIS type, typepars is a vector of 5 character strings. GRASS GIS parameters: gisBase, home, gisDbase, location, mapset.
Example parameters are included in an example script included in this package. See initGRASS help
for more info on parameters.}
}
\description{
Creates the flow network used by RHESSys
}
\author{
Will Burke
}
