% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIS_read.R
\name{GIS_read}
\alias{GIS_read}
\title{GIS_read}
\usage{
GIS_read(type, readin, typepars)
}
\arguments{
\item{type}{GIS type to be used. Currently only supports GRASS GIS.}

\item{readin}{Template file or vector containing character strings of map names.}

\item{typepars}{Parameters needed based on GIS type used. For Grass GIS type, typepars is a
vector of 5 character strings. GRASS GIS parameters: gisBase, home, gisDbase, location, mapset.
Example parameters are included in an example script included in this package. See initGRASS help
for more info on parameters.}
}
\description{
Read in GIS maps. Returns a spatial data frame containing all maps. 6/6/17.
}
\author{
Will Burke
}
