% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_flow.R
\name{read_in_flow}
\alias{read_in_flow}
\title{read_in_flow}
\usage{
read_in_flow(input_file)
}
\arguments{
\item{input_file}{Flow table to be read in.}
}
\description{
Read in flow table file and return a R list.
}
\author{
Daniel Nash
}
