% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHESSysPreprocess.R
\name{RHESSysPreprocess}
\alias{RHESSysPreprocess}
\title{RHESSysPreprocess}
\usage{
RHESSysPreprocess(template, name, type = "GRASS", typepars,
  overwrite = FALSE, asprules = NULL, streams = NULL, roads = NULL,
  impervious = NULL, roofs = NULL, wrapper = TRUE)
}
\arguments{
\item{template}{Template file used to generate worldfile for RHESSys. Generic strucutre is:
<state variable> <operator> <value/map>. Levels are difined by lines led by "_", structured
<levelname> <map> <count>. Whitespace and tabs are ignored.  Maps referred to must be supplied
by your chosen method of data input(GRASS or raster), set using the "type" arguement.}

\item{name}{The base name (and potentially, path as well) to be used for your ouput files.
This will create a world file called "<name>.world", and a flow table called "<name>.flow".}

\item{type}{Input file type to be used.  Default is "GRASS" which will attempt to autodetect the
version of GRASS GIS being used (6.x or 7.x).  GRASS GIS type can also be set explicitly to "GRASS6" or "GRASS7".
"Raster" type will use rasters in GeoTiff or equivalent format (see Raster package), with file names
matching those indicated in the template.}

\item{typepars}{Parameters needed based on input data type used. For GRASS GIS type, typepars is a
vector of 5 character strings. GRASS GIS parameters: gisBase, home, gisDbase, location, mapset.
Example parameters are included in an example script included in this package. See initGRASS help
for more info on parameters.  If using raster type, typepars should be a string indicating the path to a folder
containing the raster files that are referenced by the template. See help for the Raster function (in the Raster package)
for additional information and supported filetypes.}

\item{overwrite}{Overwrite existing worldfile. FALSE is default and prompts a menu if worldfile already exists.}

\item{asprules}{The path and filename to the rules file.  Using this argument enables aspatial patches.}

\item{wrapper}{wrapper is an internal arguement that indivates to world_gen if it should output the required data to
memory(if being run in congunction with CreateFlownet), or to files, if being fun separately.}
}
\description{
Preforms RHESSys Preprocessing, comprise of two main steps: generating a world file, and generating a flow table.
World file is generated via world_gen.R, and flow table is generated via CreateFlownet.R.
}
\seealso{
\code{\link{initGRASS}}, \code{\link{readRAST}}, \code{\link{Raster}}
}
