% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_input_gen.R
\name{spatial_input_gen}
\alias{spatial_input_gen}
\title{spatial_input_gen}
\usage{
spatial_input_gen(
  name,
  DEM,
  GRASS_path,
  threshold,
  basin = NULL,
  patch = NULL,
  define_watershed = NULL,
  easting = NULL,
  northing = NULL,
  gisDbase = NULL,
  location = NULL,
  mapset = NULL
)
}
\arguments{
\item{name}{Name of your project}

\item{DEM}{name of DEM map}

\item{GRASS_path}{Path to grass isntall location}

\item{threshold}{threshold for upstream patches to make a stream}

\item{basin}{optional input basin map}

\item{patch}{optional method to define patches}

\item{define_watershed}{Optional method to define watershed outline}

\item{easting}{easting}

\item{northing}{=northing}

\item{gisDbase}{gisDbase location}

\item{location}{GRASS GIS location}

\item{mapset}{GRASS GIS mapset}
}
\description{
Generate the spatial inputs for RHESSys and RHESSysPreprocessing
}
\author{
Will Burke
}
