% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world_gen.R
\name{world_gen}
\alias{world_gen}
\title{World Gen}
\usage{
world_gen(template, worldfile, type = "GRASS", typepars, overwrite = NULL)
}
\arguments{
\item{template}{Template file used to generate worldfile for RHESSys. Generic strucutre is:
<state variable> <operator> <value/map>. Levels are difined by lines led by "_", structured
<levelname> <map> <count>. Whitespace and tabs are ignored.}

\item{worldfile}{Name and path of worldfile to be created.}

\item{type}{GIS type to be used. Currently only supports GRASS GIS.}

\item{typepars}{Parameters needed based on GIS type used. For Grass GIS type, typepars is a
vector of 5 character strings. GRASS GIS parameters: gisBase, home, gisDbase, location, mapset.
Example parameters are included in an example script included in this package. See initGRASS help
for more info on parameters.}

\item{overwrite}{Overwrite existing worldfile. Overwrites on any non-NULL arguement.}
}
\description{
Generates world files for use in RHESSys based on input template and maps,
currently includes functionality for GRASS GIS, and works on both unix and windows. 5/16/17.
}
\seealso{
\code{\link{initGRASS}}, \code{\link{readRAST}}
}
\author{
Will Burke
}
