% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world_gen.R
\name{world_gen}
\alias{world_gen}
\title{World Gen}
\usage{
world_gen(
  template,
  worldfile,
  map_dir,
  overwrite = FALSE,
  header = FALSE,
  unique_strata_ID = TRUE,
  asprules = NULL,
  seq_patch_IDs = FALSE,
  output_patch_map = FALSE,
  convert_aspect = TRUE,
  wrapper = FALSE
)
}
\arguments{
\item{template}{Template file used to generate worldfile for RHESSys. Generic strucutre is:
<state variable> <operator> <value/map>. Levels are difined by lines led by "_", structured
<levelname> <map> <count>. Whitespace and tabs are ignored.  Maps referred to must be supplied
by your chosen method of data input(raster).}

\item{worldfile}{Name and path of worldfile to be created.}

\item{map_dir}{Character string indicating the path to a folder containing the raster files that are referenced by the template.}

\item{overwrite}{Overwrite existing worldfile. FALSE is default and prompts a menu if worldfile already exists.}

\item{header}{TRUE/FALSE flag for the creation of a header file. Will have same name (and location) as "name" argument, but with ".hdr" suffix.}

\item{unique_strata_ID}{Takes input map or value for canopy strata ID and appends either a 1 or 2 dpending on which canopy it is. Defaults to TRUE.}

\item{asprules}{The path and filename to the rules file.  Using this argument enables aspatial patches.}

\item{seq_patch_IDs}{TRUE/FALSE should patch map IDs be overwritten with sequential integers.}

\item{output_patch_map}{TRUE/FALSE should the new patch map with sequential IDs be output to file.}

\item{convert_aspect}{TRUE/FALSE if the input aspect map data should be converted from GRASS GIS format (CCW EAST==0) to the 'normal'
format needed for RHESSys (CW NORTH==0). Defaults to TRUE.}

\item{wrapper}{internal argument to track if being run as all-in-one}
}
\description{
Generates world files for use in RHESSys based on input template and maps.
}
\seealso{
\code{\link{raster}}
}
\author{
Will Burke
}
